AddApiDashBoard

+++ src/main/java/com/swp391/project/service/impl/ProjectServiceImp.java
@@ -22,7 +22,7 @@ public interface ProjectServiceImp {
 
     boolean updateProjectByStatus(int projectId, int userId, String status);
 
-    int findTotalProjectByStatus(String status);
+    long findTotalProjectByStatus(String status);
 
     ProjectDTO findById(int id);
